+++ CHANGES.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- LoRA
+- transformer weights dtype
+
+### Added
+
 - Resolve HF tokenizer model max length when undeclared.
 
 ### Changes